Watch for trap in GDB
GDB doesn't appear to play nice with the riscv-tests, especially the VM
ones, so when a breakpoint is reached while inside virtual memory it
doesn't know where to look for instruction information and can only say
that a breakpoint trap occurred. As a result, the runtime has been
updated to include watching for the word "trap" in the GDB log to signal
the end of the test.